+++ helpers.php
@@ -126,30 +126,3 @@ function excerpt($str, $limit = 55)
 
     return $rc;
 }
-
-/**
- * Removes the `DOCUMENT_ROOT` from the provided path.
- *
- * Note: Because this function is usually used to create URL path from server path, the directory
- * separator '\' is replaced by '/'.
- *
- * @param string $pathname
- *
- * @return string
- */
-function strip_root($pathname)
-{
-    $root = rtrim($_SERVER['DOCUMENT_ROOT'], DIRECTORY_SEPARATOR);
-    $root = strtr($root, DIRECTORY_SEPARATOR == '/' ? '\\' : '/', DIRECTORY_SEPARATOR);
-    $pathname = strtr($pathname, DIRECTORY_SEPARATOR == '/' ? '\\' : '/', DIRECTORY_SEPARATOR);
-
-    if ($root && strpos($pathname, $root) === 0) {
-        $pathname = substr($pathname, strlen($root));
-    }
-
-    if (DIRECTORY_SEPARATOR != '/') {
-        $pathname = strtr($pathname, DIRECTORY_SEPARATOR, '/');
-    }
-
-    return $pathname;
-}